refactor: ResultModal 컴포넌트 함수형으로 변경
